refactor(auth): delete the legacy passkey promotion path
Native and web clients now use Better Auth's transactional registration hook,
so remove finalize-passkey, passkey ownership polling, the promotion SQL, route,
OpenAPI surface, tests, and current-flow documentation.

Keep the durable account-upgrade event by emitting it from afterVerification,
the code that now performs the in-place upgrade. The event is committed in the
same registration transaction as the passkey and session instead of being owned
by a second endpoint.

Also delete the unused anonymous data migration helper and tests. Upgrade in
place preserves the user id, so no rows move between accounts.
